Add [attempts tab] profile
