+++ src/components/Task/Results/Results.tsx
@@ -21,212 +21,17 @@ import {
   errorNotification,
   newNotification,
 } from '@utils/notificationFunctions';
-
-const refactorAttempt = (
-  attempt: IAttemptDisplay,
-  locale: ILocale
-): any => ({
-  ...attempt,
-  result: {
-    display: (
-      <div
-        style={{
-          color:
-            attempt.status.spec == 2
-              ? attempt.verdict.verdict.spec == 0
-                ? 'var(--positive)'
-                : 'var(--negative)'
-              : 'black',
-        }}
-      >
-        {attempt.status.spec == 2
-          ? attempt.verdict.verdict.shortText +
-            ' #' +
-            (attempt.verdict.test + 1).toString()
-          : locale.attempt.statuses[attempt.status.spec]}
-      </div>
-    ),
-    value:
-      attempt.status.spec == 2
-        ? attempt.verdict.verdict.spec
-        : attempt.status.spec - 10,
-  },
-  date: {
-    display: <>{getLocalDate(attempt.date)}</>,
-    value: new Date(attempt.date).getTime(),
-  },
-  language: {
-    display: <>{attempt.language.name}</>,
-    value: attempt.language,
-  },
-});
-
-const initialColumns = (locale: ILocale): ITableColumn[] => [
-  {
-    label: locale.attempt.date,
-    key: 'date',
-    sortable: true,
-    sortFunction: (a: any, b: any) =>
-      a.date.value > b.date.value
-        ? -1
-        : a.date.value == b.date.value
-        ? 0
-        : 1,
-    sorted: -1,
-    allowMiddleState: false,
-    hidable: false,
-    hidden: false,
-    size: 5,
-  },
-  {
-    label: locale.attempt.language,
-    key: 'language',
-    sortable: false,
-    sortFunction: (a: any, b: any) => 0,
-    sorted: 0,
-    allowMiddleState: true,
-    hidable: false,
-    hidden: false,
-    size: 5,
-  },
-  {
-    label: locale.attempt.result,
-    key: 'result',
-    sortable: false,
-    sortFunction: (a: any, b: any) => 0,
-    sorted: 0,
-    allowMiddleState: true,
-    hidable: false,
-    hidden: false,
-    size: 5,
-  },
-];
-
-const defaultOnPage = 10;
-
-interface PagerResponse {
-  data: IAttemptDisplay[];
-  total: number;
-}
-
-interface TableData {
-  data: any[];
-  total: number;
-}
+import AttemptList from '@ui/AttemptList/AttemptList';
 
 const Results: FC<{ spec: string; activeTab: string }> = ({
   spec,
   activeTab,
 }) => {
-  const { locale } = useLocale();
-  const { refreshAccess } = useUser();
-
-  const columns: ITableColumn[] = useMemo(
-    () => initialColumns(locale),
-    [locale]
-  );
-
-  const [needRefetch, setNeedRefetch] = useState(true);
-  const [tableData, setTableData] = useState<TableData>({
-    data: [],
-    total: 0,
-  });
-
-  const [searchParams, setSearchParams] = useState<BaseSearch>({
-    pager: {
-      skip: 0,
-      limit: defaultOnPage,
-    },
-    sort_by: [{ field: 'date', order: -1 }],
-    search_params: {
-      search: '',
-      keys: [],
-    },
-  });
-
-  const processData = useCallback(
-    (response: PagerResponse): TableData => ({
-      data: response.data.map((item) =>
-        refactorAttempt(item, locale)
-      ),
-      total: response.total,
-    }),
-    [locale]
-  );
-
-  const onError = useCallback(
-    (res: any) => {
-      if (refreshAccess() == 2) {
-        setTableData({ data: [], total: 0 });
-        const id = newNotification({});
-        errorNotification({
-          id,
-          title: locale.notify.errors.unauthorized,
-          autoClose: 10000,
-        });
-        setNeedRefetch(false);
-      } else {
-        setNeedRefetch(true);
-      }
-    },
-    [locale.notify.errors.unauthorized, refreshAccess]
-  );
-
-  const { data, loading, refetch } = useRequest<
-    BaseSearch,
-    PagerResponse,
-    TableData
-  >(
-    `task/attempts/${spec}`,
-    'POST',
-    searchParams,
-    processData,
-    undefined,
-    onError
-  );
-
-  useEffect(() => {
-    if (data) {
-      setTableData(data);
-    }
-  }, [data]);
-
-  useEffect(() => {
-    refetch();
-    if (activeTab == 'results') {
-      const id = setInterval(() => {
-        if (needRefetch) refetch(false);
-      }, 2000);
-      return () => {
-        clearInterval(id);
-      };
-    }
-  }, [needRefetch, refetch, activeTab]);
-
   return (
-    <div>
-      <Table
-        columns={columns}
-        rows={tableData.data}
-        total={tableData.total}
-        loading={loading}
-        setSearchParams={setSearchParams}
-        searchParams={searchParams}
-        classNames={{
-          wrapper: tableStyles.wrapper,
-          table: tableStyles.table,
-          author: tableStyles.author,
-          grade: tableStyles.grade,
-          verdict: tableStyles.verdict,
-          headerCell: tableStyles.headerCell,
-          cell: tableStyles.cell,
-          even: tableStyles.even,
-          odd: tableStyles.odd,
-        }}
-        defaultOnPage={defaultOnPage}
-        onPage={[5, 10]}
-      />
-    </div>
+    <AttemptList
+      url={`task/attempts/${spec}`}
+      activeTab={activeTab === 'results'}
+    />
   );
 };
