+++ src/components/ui/Table/table.module.css
@@ -1,6 +1,4 @@
 .wrapper {
-  width: 80vw;
-  margin: var(--spacer-s) 10vw 10vh 10vw;
   height: 100%;
   font-size: var(--font-size-l);
   display: flex;
